# Improve performance of DataFrame binary comparison operations