… src/Microsoft.Data.Analysis/Computations/Arithmetics.cs
… src/Microsoft.Data.Analysis/Strings.Designer.cs
… src/Microsoft.Data.Analysis/VBufferDataFrameColumn.cs
… test/Microsoft.Data.Analysis.PerformanceTests/PerformanceTests.cs