Goal of this PR is to improve performance of comparison operation and is the next step in aligning datafrane arithmetic API with the new TensorPrimitives API

In DataFrame 0.20.1:

|                          Method |     Mean |    Error |   StdDev |
|-------------------------------- |---------:|---------:|---------:|
|   ElementwiseEquals_Int32_Int32 | 38.00 ms | 0.145 ms | 0.121 ms |
|   ElementwiseEquals_Int16_Int16 | 39.55 ms | 0.291 ms | 0.258 ms |
| ElementwiseEquals_Double_Double | 40.28 ms | 0.367 ms | 0.343 ms |
|   ElementwiseEquals_Float_Float | 41.18 ms | 0.805 ms | 1.074 ms |

After this PR:

|                          Method |      Mean |     Error |    StdDev |
|-------------------------------- |----------:|----------:|----------:|
|   ElementwiseEquals_Int32_Int32 |  1.171 ms | 0.0228 ms | 0.0263 ms |
|   ElementwiseEquals_Int16_Int16 |  1.090 ms | 0.0569 ms | 0.0475 ms |
| ElementwiseEquals_Double_Double |  1.388 ms | 0.0264 ms | 0.0247 ms |
|   ElementwiseEquals_Float_Float |  1.250 ms | 0.0215 ms | 0.0190 ms |``

Other comparison operations shows the same boost in performance